Remade the "Redirect To Website Of User Language.js" JavaScript file. It is on the root website folder in the "JavaScript/Function" folder. Renamed it to "Language Redirector.js". Coded it using an array of English website names and a property (dictionary) of Portuguese website names with English website names as keys. Used the Fetch API to read the two files in the root website folder in the "Texts/Websites List" folder. And JSON.parse to parse the files into array and property (dictionary). Now every website needs to be only inside of the English website folder for the Redirect to work. The "European Portuguese" (PT-PT) language will no more use a different URL and folder name, it will be the same as the "PT-BR" language. Created two new functions, "English_Language_Check", and "Portuguese_Language_Check". To check if the user language is English or Portuguese. If the website is the index one, redirect to the user language website. If the website title is in English and the user language is Portuguese, redirect to the Portuguese website. If the website title is in Portuguese and the user language is English, redirect to the English website. Created a new function called "Redirect" to redirect to the user language website and also add the URL addon.
Created a way to hide the big space in computer mode in the "Show Hide.js" JavaScript file on the root website folder in the "JavaScript/Function" folder.
By creating a div with br elements with the id "computer_buttons_bar_space" in "Body.html" RainTPL template.
Hiding it when clicking the hide button, showing it when clicking on the show button, on the JavaScript hide and show functions, respectively.

Updated "Watch History" website to add new watched things and change things being watched.
Updated "The Life of Littletato", "SpaceLiving", "The Secret of the Crystals", and "The Story of the Bulkan Siblings" to update style and elements.