I greatly improved the JavaScript code of the "Story.js" file, improving its log text shown on the console. I joined the "Define_Chapter" and "Open_Chapter" texts together, to show all in one go. I also added a way to scroll to the "write chapter" text area if it exists and added the page scrolling log text.
Updated websites:
I updated the "Stake2" website to fix and update the profile pictures of the Digital Identities and some elements of the website.
I updated the "SpaceLiving", "The Life of Littletato", and "The Secret of the Crystals" websites to update the chapters, chapter buttons, style, and "Other stories" tab, and fixed some variables.
I updated the "Watch History" website to update the watched things, titles of the website, descriptions, tab titles, and layout.